requirements.txt
# Guardian Dashboard - Clean Dependencies
streamlit>=1.28.0
firebase-admin>=6.2.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.17.0
requests>=2.31.0
google-cloud-firestore>=2.11.0
